# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: ÖzelGüvenlik.Online API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /listings:
    get:
      operationId: getListings
      tags: [listings]
      summary: Get all listings
      parameters:
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
        - name: city
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
        - name: featured
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: List of job listings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListingsResponse"

    post:
      operationId: createListing
      tags: [listings]
      summary: Create a new listing
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ListingInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Listing"
  /listings/{id}:
    get:
      operationId: getListing
      tags: [listings]
      summary: Get listing by id
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Listing
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Listing"
        "404":
          description: Not found

    patch:
      operationId: updateListing
      tags: [listings]
      summary: Update a listing
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ListingUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Listing"

    delete:
      operationId: deleteListing
      tags: [listings]
      summary: Delete a listing
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /listings/{id}/like:
    post:
      operationId: toggleListingLike
      tags: [listings]
      summary: Toggle like on a listing
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Like toggled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LikeResponse"
  /listings/{id}/favorite:
    post:
      operationId: toggleListingFavorite
      tags: [listings]
      summary: Toggle favorite on a listing
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Favorite toggled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FavoriteResponse"
  /listings/{id}/view:
    post:
      operationId: incrementListingView
      tags: [listings]
      summary: Increment view count
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: View incremented
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
  /admin/listings/{id}/approve:
    post:
      operationId: approveListing
      tags: [admin]
      summary: Approve a listing
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Approved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /admin/listings/{id}/feature:
    post:
      operationId: featureListing
      tags: [admin]
      summary: Toggle featured status of a listing
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Featured toggled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /admin/listings/{id}/fake-likes:
    post:
      operationId: addFakeLikes
      tags: [admin]
      summary: Add fake likes to a listing (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FakeLikesInput"
      responses:
        "200":
          description: Fake likes added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
components:
  schemas:
...
    SuccessResponse:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
        message:
          type: ["string", "null"]
...
    Listing:
      type: object
      required: [id, title, company, city, status, viewCount, likeCount, createdAt]
      properties:
        id:
          type: integer
        title:
          type: string
        company:
          type: string
        city:
          type: string
        salary:
          type: ["string", "null"]
        workType:
          type: string
        description:
          type: ["string", "null"]
        requirements:
          type: ["string", "null"]
        status:
          type: string
        viewCount:
          type: integer
        likeCount:
          type: integer
        isFeatured:
          type: boolean
        applyUrl:
          type: ["string", "null"]
        companyLogoUrl:
          type: ["string", "null"]
        authorId:
          type: ["integer", "null"]
        authorUsername:
          type: ["string", "null"]
        isLikedByMe:
          type: boolean
        isFavoritedByMe:
          type: boolean
        createdAt:
          type: string
    ListingInput:
      type: object
      required: [title, company, city]
      properties:
        title:
          type: string
        company:
          type: string
        city:
          type: string
        salary:
          type: ["string", "null"]
        workType:
          type: string
        description:
          type: ["string", "null"]
        requirements:
          type: ["string", "null"]
        applyUrl:
          type: ["string", "null"]
        companyLogoUrl:
          type: ["string", "null"]
    ListingUpdate:
      type: object
      properties:
        title:
          type: string
        company:
          type: string
        city:
          type: string
        salary:
          type: ["string", "null"]
        workType:
          type: string
        description:
          type: ["string", "null"]
        requirements:
          type: ["string", "null"]
        status:
          type: string
        applyUrl:
          type: ["string", "null"]
        isFeatured:
          type: boolean
    ListingsResponse:
      type: object
      required: [listings, total, page, limit]
      properties:
        listings:
          type: array
          items:
            $ref: "#/components/schemas/Listing"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
...
    LikeResponse:
      type: object
      required: [liked, likeCount]
      properties:
        liked:
          type: boolean
        likeCount:
          type: integer
    FavoriteResponse:
      type: object
      required: [favorited]
      properties:
        favorited:
          type: boolean
...
    FakeLikesInput:
      type: object
      required: [count]
      properties:
        count:
          type: integer
...